change increments to bigIncrements

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,7 +14,7 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('firstName',30);
             $table->string('lastName',30);
             $table->string('email',50)->unique();
@@ -25,14 +25,14 @@ class CreateUsersTable extends Migration
             $table->rememberToken();
             $table->timestamps();
             $table->softDeletes();
-            $table->integer('created_by')->nullable();
+            $table->bigInteger('created_by')->nullable();
             //$table->foreign('created_by')->references('id')->on('users');//->onUpdate('cascade')->onDelete('cascade');  
-            $table->integer('updated_by')->nullable();
+            $table->bigInteger('updated_by')->nullable();
             //$table->foreign('updated_by')->references('id')->on('users');//->onUpdate('cascade')->onDelete('cascade');   
-            $table->integer('deleted_by')->nullable();
+            $table->bigInteger('deleted_by')->nullable();
             //$table->foreign('deleted_by')->references('id')->on('users');//->onUpdate('cascade')->onDelete('cascade');  
             $table->timestamp("restored_at")->nullable();
-            $table->integer('restored_by')->unsigned()->nullable();
+            $table->bigInteger('restored_by')->unsigned()->nullable();
             // $table->foreign('restored_by')->references('id')->on('users'); 
 
         });
