+++ database/migrations/2018_04_26_173721_create_ramadans_table.php
@@ -14,7 +14,7 @@ class CreateRamadansTable extends Migration
     public function up()
     {
         Schema::create('ramadans', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->date('date');
             $table->string('sehrTime',4);
             $table->string('fajrTime',4)->nullable(true);
@@ -24,14 +24,14 @@ class CreateRamadansTable extends Migration
             $table->unique(['date', 'areaCode']);
             $table->timestamps();
             $table->softDeletes();
-            $table->integer('created_by')->unsigned()->nullable();
+            $table->bigInteger('created_by')->unsigned()->nullable();
             $table->foreign('created_by')->references('id')->on('users');  
-            $table->integer('updated_by')->unsigned()->nullable();
+            $table->bigInteger('updated_by')->unsigned()->nullable();
             $table->foreign('updated_by')->references('id')->on('users');   
-            $table->integer('deleted_by')->unsigned()->nullable();
+            $table->bigInteger('deleted_by')->unsigned()->nullable();
             $table->foreign('deleted_by')->references('id')->on('users'); 
             $table->timestamp("restored_at")->nullable();
-            $table->integer('restored_by')->unsigned()->nullable();
+            $table->bigInteger('restored_by')->unsigned()->nullable();
             $table->foreign('restored_by')->references('id')->on('users'); 
         });
     }
