+++ database/migrations/2018_04_29_170834_create_dayflags_table.php
@@ -14,20 +14,20 @@ class CreateDayflagsTable extends Migration
     public function up()
     {
         Schema::create('dayflags', function (Blueprint $table) {
-            $table->integer('flag')->primary();
+            $table->bigInteger('flag')->primary();
             $table->string('name_en');
             $table->string('name_bn');
             $table->integer('display_order');
             $table->timestamps();
             $table->softDeletes();
-            $table->integer('created_by')->unsigned()->nullable();
+            $table->bigInteger('created_by')->unsigned()->nullable();
             $table->foreign('created_by')->references('id')->on('users');  
-            $table->integer('updated_by')->unsigned()->nullable();
+            $table->bigInteger('updated_by')->unsigned()->nullable();
             $table->foreign('updated_by')->references('id')->on('users');   
-            $table->integer('deleted_by')->unsigned()->nullable();
+            $table->bigInteger('deleted_by')->unsigned()->nullable();
             $table->foreign('deleted_by')->references('id')->on('users'); 
             $table->timestamp("restored_at")->nullable();
-            $table->integer('restored_by')->unsigned()->nullable();
+            $table->bigInteger('restored_by')->unsigned()->nullable();
             $table->foreign('restored_by')->references('id')->on('users');   
         });
     }
