+++ database/migrations/2018_04_30_164751_create_days_table.php
@@ -14,7 +14,7 @@ class CreateDaysTable extends Migration
     public function up()
     {
         Schema::create('days', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('photoFileName')->nullable();
             $table->string('photoCaptionEn')->nullable();
             $table->string('photoCaptionBn')->nullable();
@@ -30,20 +30,20 @@ class CreateDaysTable extends Migration
             $table->string('dayCategory',3)->nullable();
             $table->string('religionCode',3)->nullable();
             $table->foreign('religionCode')->references('code')->on('religions')->onUpdate('cascade');
-            $table->integer('dayFlag')->nullable();  
+            $table->bigInteger('dayFlag')->nullable();  
             //$table->foreign('dayFlag')->references('flag')->on('dayflags')->onUpdate('cascade'); 
             $table->string('holidayCode',2)->nullable();
             $table->foreign('holidayCode')->references('code')->on('holidaytypes')->onUpdate('cascade');   
             $table->timestamps();
             $table->softDeletes();
-            $table->integer('created_by')->unsigned()->nullable();
+            $table->bigInteger('created_by')->unsigned()->nullable();
             $table->foreign('created_by')->references('id')->on('users');  
-            $table->integer('updated_by')->unsigned()->nullable();
+            $table->bigInteger('updated_by')->unsigned()->nullable();
             $table->foreign('updated_by')->references('id')->on('users');   
-            $table->integer('deleted_by')->unsigned()->nullable();
+            $table->bigInteger('deleted_by')->unsigned()->nullable();
             $table->foreign('deleted_by')->references('id')->on('users'); 
             $table->timestamp("restored_at")->nullable();
-            $table->integer('restored_by')->unsigned()->nullable();
+            $table->bigInteger('restored_by')->unsigned()->nullable();
             $table->foreign('restored_by')->references('id')->on('users');   
         });
     }
